fix disabled button and move code to different file
